requirements.txt
imutils==0.5.4
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
setuptools==69.5.1
wheel==0.43.0
